+++ src/snowflake/snowpark/modin/plugin/compiler/snowflake_query_compiler.py
@@ -59,7 +59,12 @@ from pandas.api.types import (
     is_string_dtype,
 )
 from pandas.core.dtypes.base import ExtensionDtype
-from pandas.core.dtypes.common import is_dict_like, is_list_like, pandas_dtype
+from pandas.core.dtypes.common import (
+    is_dict_like,
+    is_list_like,
+    is_timedelta64_dtype,
+    pandas_dtype,
+)
 from pandas.core.indexes.base import ensure_index
 from pandas.io.formats.format import format_percentiles
 from pandas.io.formats.printing import PrettyDict
@@ -8807,6 +8812,10 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
                 labels, include_index=False
             )
         )
+        data_column_snowpark_pandas_types = [
+            TimedeltaType() if is_timedelta64_dtype(type) else None
+            for type in col_dtypes_map.values()
+        ]
         for ids, label in zip(col_ids, labels):
             for id in ids:
                 to_dtype = col_dtypes_map[label]
@@ -8826,7 +8835,8 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
 
         return SnowflakeQueryCompiler(
             self._modin_frame.update_snowflake_quoted_identifiers_with_expressions(
-                astype_mapping
+                quoted_identifier_to_column_map=astype_mapping,
+                data_column_snowpark_pandas_types=data_column_snowpark_pandas_types,
             ).frame
         )